[fix] [refactor] [license-table] Fix inconsistent UI and refactor some styles to be modular.
